Rename directory from utils/server to  utils/db
