+++ src/app/_component/Header.tsx
@@ -1,6 +1,6 @@
 import { GetStartedButton } from "@/components";
 import { SiteLogo } from "@/components/icons";
-import { checkAuthClient } from "@/utils/server/auth";
+import { checkAuthClient } from "@/utils/db/auth";
 
 export const Header = async () => {
   const isLoggedIn = await checkAuthClient();
