+++ src/app/auth/_components/SignUpForm.tsx
@@ -14,7 +14,7 @@ import {
   FieldGroup,
   FieldLabel,
 } from "@/components/ui/field";
-import { loginGoogle, signup } from "@/utils/server/auth";
+import { loginGoogle, signup } from "@/utils/db/auth";
 import { Eye, EyeOff } from "lucide-react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
