+++ src/app/auth/callback/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@/utils/supabase/server";
-import { getProfileByUserId } from "@/utils/server/profile";
+import { getProfileByUserId } from "@/utils/db/profile";
 
 export async function GET(request: Request) {
   const { searchParams, origin } = new URL(request.url);
