+++ src/app/auth/forgot-password/_component/ResetPassWordForm.tsx
@@ -13,7 +13,7 @@ import {
   FieldGroup,
   FieldLabel,
 } from "@/components/ui/field";
-import { resetPassWord } from "@/utils/server/auth";
+import { resetPassWord } from "@/utils/db/auth";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { Controller, useForm } from "react-hook-form";
