+++ src/app/dashboard/_components/ReguralrFoodListItem.tsx
@@ -1,6 +1,6 @@
 import { List, Loading } from "@/components";
 import { SelectregularFood } from "@/db/schema";
-import { addMealRecord } from "@/utils/server/mealRecords";
+import { addMealRecord } from "@/utils/db/mealRecords";
 import { createJstDate, formatTime, formatYYMMDD } from "@/utils/format/date";
 import { historieskeys, mealRecordkeys } from "@/utils/tanstack";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
