+++ src/app/dashboard/histories/_components/HistoryList.tsx
@@ -7,7 +7,7 @@ import { Button } from "@/components/ui";
 import {
   DailyKcalSummaryResponse,
   fetchDailyKcalSummary,
-} from "@/utils/server/history";
+} from "@/utils/db/history";
 import { historieskeys, TErrCodes } from "@/utils/tanstack";
 import { Loading } from "@/components";
 import { FetchErrorMessage } from "@/app/dashboard/_components";
