+++ src/app/dashboard/histories/page.tsx
@@ -1,6 +1,6 @@
 import { HistoryList } from "@/app/dashboard/histories/_components";
-import { fetchDailyKcalSummary } from "@/utils/server/history";
-import { getUserId } from "@/utils/server/auth";
+import { fetchDailyKcalSummary } from "@/utils/db/history";
+import { getUserId } from "@/utils/db/auth";
 import { getQueryClient, historieskeys } from "@/utils/tanstack";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
