+++ src/app/dashboard/target-kcal-plans/_component/TargetKcalActionMenu.tsx
@@ -2,7 +2,7 @@ import { TargetKcalPlanForm } from "@/app/dashboard/target-kcal-plans/_component
 import { Loading } from "@/components";
 import { SelectTargetKcalPlansRecord } from "@/db/schema";
 import { useModalControl } from "@/hooks";
-import { deleteTargetKcal } from "@/utils/server/targetKcal";
+import { deleteTargetKcal } from "@/utils/db/targetKcal";
 import { TargetKcalkeys } from "@/utils/tanstack";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { EllipsisVertical, Pencil, Trash2, X } from "lucide-react";
