+++ src/app/dashboard/target-kcal-plans/page.tsx
@@ -1,6 +1,6 @@
 import { TargetKcalSection } from "@/app/dashboard/target-kcal-plans/_component";
-import { fetchUserTargetKcal } from "@/utils/server/targetKcal";
-import { getUserId } from "@/utils/server/auth";
+import { fetchUserTargetKcal } from "@/utils/db/targetKcal";
+import { getUserId } from "@/utils/db/auth";
 import { getQueryClient, TargetKcalkeys } from "@/utils/tanstack";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
