+++ src/app/setup/_components/TargetKcalForm.tsx
@@ -13,7 +13,7 @@ import {
   FieldGroup,
   FieldLabel,
 } from "@/components/ui/field";
-import { createTargetKcal } from "@/utils/server/setup";
+import { createTargetKcal } from "@/utils/db/setup";
 import { formatYYMMDD } from "@/utils/format/date";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
