+++ src/app/setup/step-2-target-kcal/page.tsx
@@ -1,5 +1,5 @@
 import { TargetKcalForm } from "@/app/setup/_components";
-import { getUserId } from "@/utils/server/auth";
+import { getUserId } from "@/utils/db/auth";
 import React from "react";
 
 export default async function SetFirstTargetKcal() {
